% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemTrajR-package.R
\docType{package}
\name{MemTrajR-package}
\alias{MemTrajR-package}
\alias{MemTrajR}
\title{MemTrajR: Quantify Micron-Scale Rigid Domain Dynamics and Interactions}
\description{
MemTrajR provides tools for analysing 2D and 3D membrane domain
trajectories from single-particle tracking data on spherical vesicles.
}
\section{MSD analysis}{

Use \code{\link[=compute_msd_2d]{compute_msd_2d()}} and \code{\link[=compute_msd_3d]{compute_msd_3d()}} to compute mean squared
displacement curves, \code{\link[=fit_msd]{fit_msd()}} to fit them, and \code{\link[=analyze_domain]{analyze_domain()}}
as a convenient all-in-one wrapper.
}

\section{Separation and interaction potential}{

Use \code{\link[=compute_geodesic_separation]{compute_geodesic_separation()}} to compute frame-by-frame separations,
\code{\link[=bin_separations]{bin_separations()}} to histogram them, \code{\link[=estimate_interaction_potential]{estimate_interaction_potential()}}
for Boltzmann analysis, \code{\link[=compute_mssep]{compute_mssep()}} and \code{\link[=fit_mssep]{fit_mssep()}} for the mean
square separation increment, and \code{\link[=analyze_separation]{analyze_separation()}} as an all-in-one
wrapper.
}

\author{
\strong{Maintainer}: Anthony Prempeh \email{anthonynaprempeh@gmail.com}

}
